+++ src/utils/globals.hpp
@@ -6,6 +6,7 @@ namespace ng::globals
 {
 	inline bool g_isEditorIDAPILoaded = false;
 	inline bool g_isBetterEditLoaded = false;
+	inline bool g_isImprovedGroupViewLoaded = false;
 
 	inline std::string g_buildHelperRawNameFormat = "";
 	inline std::string g_buildHelperNameFormat = "";